Fix data logging type casting

+++ worker/include/Logger.hpp
@@ -345,7 +345,7 @@ public:
 		  { \
 		  	if (bufferDataLen != 0) \
 		  	{ \
-		  		Logger::channel->SendLog(Logger::buffer, bufferDataLen); \
+		  		Logger::channel->SendLog(Logger::buffer, static_cast<uint32_t>(bufferDataLen)); \
 		  		bufferDataLen = 0; \
 		  	} \
 		    int loggerWritten = std::snprintf(Logger::buffer + bufferDataLen, Logger::bufferSize, "X%06X ", static_cast<unsigned int>(i)); \
@@ -355,7 +355,7 @@ public:
 		  bufferDataLen += loggerWritten; \
 		} \
 		if (bufferDataLen != 0) \
-			Logger::channel->SendLog(Logger::buffer, bufferDataLen); \
+			Logger::channel->SendLog(Logger::buffer, static_cast<uint32_t>(bufferDataLen)); \
 	} \
 	while (false)
